contineud cleanup
