+++ crates/bitwarden-policies/src/lib.rs
@@ -5,15 +5,17 @@ uniffi::setup_scaffolding!();
 #[cfg(feature = "uniffi")]
 mod uniffi_support;
 
+mod enriched_policy;
 mod enriched_policy_type;
-pub mod filter;
 mod models;
 mod policy_client;
+pub mod policy_definition;
 pub mod policy_definitions;
 mod policy_type;
 
+pub use enriched_policy::EnrichedPolicy;
 pub use enriched_policy_type::EnrichedPolicyType;
-pub use filter::PolicyDefinition;
+pub use policy_definition::PolicyDefinition;
 pub use models::{OrganizationUserPolicyContext, PolicyView};
 pub use policy_client::{PoliciesClientExt, PolicyClient};
 pub use policy_definitions::*;
